requirements.txt
matplotlib==3.8.2
mplfinance==0.12.9
Pillow==10.1.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
newsapi==0.1.0
finnhub-client==1.4.10
